Add custom login page
Created our own login, logout, home page.
We need to disable the default page. Instead of that default page call our own custom page.
In security config file, override the config method with HttpSercurity as parameter.
Add configuration to call custom login page.
Invalidate session on logout and redirect it to home page.
Added configuration for url to be authenticated.